v2.0.2 Conversion to flux-based API continued..
Now roughly half of the API is flux based, and the other half is still
using nested/propogated signals, so this commit is non-functioning.  It
was simply committed as a save point before I dive into the complexities
that is the launching mechanism (which will likely require a number of
changes to the Blocks, their parents, and there will be a significant
addition of graphical components starting with 2.1.0.

In fact, this version has some of the framework in place now for those
sweeping changes. Stay tuned.