Working v6.2.0-alpha
blocks correctly spawn, match, launch, and re-fill in reliable order,
then continue to match or if no matches,  end move and if move count >
3, end turn.

Had to convert the entire business logic into C++ because QML-based
business logic just does not work properly -- even if every little
change is carefully tracked, there will be times when the JS engine will
execute a command, and then halfway through it, something else will
execute at the same time, changing the parameters or data, but the
previous command will not have received the change an thereby reports
the incorrect values based on the previous calculations, making it
impossible to reliably track the state of the game over remote
communication without sending every change over the network and then
waiting for the other player to catch up before performing the next
action.

This is a major buzz kill and will result in countless games hanging
when one user no longer responds.

Also the bandwidth requirements are too big for this type of game.

One alternative is to do an even-more-than-currently compressed data by
minifying everything down to single characters, but that is optimization
for pre-release phase, not alpha.

Unfortunately, there are limitations to how much data can be spammed
into the IRC servers, even after modfication.

The fact is that every message sent is forwarded to the #services.log
channel for tracking game outcomes etc, and I will have to impove he
services by having them handle all of the game tracking rather than
having an IRCop client sitting locally which will be somewhat difficult.

I think having all of the blocks controlled by C++ is what the original
QML designers intended anyways, so I will go with that tactic even if it
means that I have to create a whole bunch of signals that are used to
control the game board.

Luckily with quickflux, I don't have to create a new signal/slot for each
object that the signals have to pass through -- this would be close to
impossible in that situation.