v 2.3.0-alpha
Finally have a semi-reliable alpha available that has moves, offense,
defense, working blocks and fairly reliable game play.

There are periodic glitches where blocks don't behave exactly the way
that they should.  This is really only because of optimizations that
were made to decrease the overall footprint of the code by merging
instructions and storing intermediate data in memory outside of loops.

All that has to be done is to track down the points of failure for the
few situations where bugs do occur and move some of the dependencies
back into the loop.

Another issue is with the colors not updating at the correct time, which
is 99% fixed, but occasionally has unexpected color shifts. The color
shifts can be fixed by finding the correct timing for updating just the
block colors from the engine to QML