2.4.1 optimized
Moved many of the flux listeners from within the QML Block objects and
instead moved them to the ArmyBlocks QML object since its the parent of
all blocks, which removed most of the lag, leaving only a small bit of
lag from the listeners that have not been moved yet.

Ultimately, there will be no listeners in the Block QML object, only
functions that are called by ArmyBlocks.

The savings here was something like 9800% fewer calls to the Dispatch
method from flux.